flask-backend/requirements.txt
# Flask API Backend for RescueRadar
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
groq==0.4.2
supabase==1.0.4
twilio==8.5.0
qrcode[pil]==7.4.2
requests==2.31.0
werkzeug==2.3.7
postgrest==0.10.8
